refactor: merge `dom.js` related tests into single `dom.spec.js` file
